prepare to run db:update on production

+++ db/reset.php
@@ -3,7 +3,8 @@
 $envs = include realpath(__DIR__ . '/../../ksiaki_config.php');
 
 $env = getenv('APPLICATION_ENV');
-$config = empty($env) ? $envs['development'] : $envs[$env];
+if (!$env) $env = 'development';
+$GLOBALS['config'] = $envs[$env];
 
 list($host, $dbname, $user, $pass, $port) = [
   $config['DB_HOST'],
@@ -13,19 +14,28 @@ list($host, $dbname, $user, $pass, $port) = [
   $config['DB_PORT']
 ];
 
-if (!exec("mysql -h $host -P $port -u $user --password=$pass -e \"DROP DATABASE IF EXISTS $dbname\"")) {
-  echo "\nDatabase '$dbname' dropped if existed\n";
-}
+$match = '*.sql';
 
-if (!exec("mysql -h $host -P $port -u $user --password=$pass -e \"CREATE DATABASE $dbname CHARACTER SET utf8 COLLATE utf8_polish_ci\"")) {
-  echo "Database '$dbname' created\n\n";
+if (in_array($env, ['development', 'testing'])) {
+
+    if (!exec("mysql -h $host -P $port -u $user --password=$pass -e \"DROP DATABASE IF EXISTS $dbname\"")) {
+      echo "\nDatabase '$dbname' dropped if existed\n";
+    }
+
+    if (!exec("mysql -h $host -P $port -u $user --password=$pass -e \"CREATE DATABASE $dbname CHARACTER SET utf8 COLLATE utf8_polish_ci\"")) {
+      echo "Database '$dbname' created\n\n";
+    }
+
+} else {
+    echo "\nRunning in '$env' – SAFE MODE schema and data unchanged\n\n";
+    $match = '1*sp*.sql';
 }
 
 $path = realpath(__DIR__ . "/migrations");
 $files = scandir($path);
 
 foreach ($files as $file) {
-    if (!fnmatch('*.sql', $file)) continue;
+    if (!fnmatch($match, $file)) continue;
 
     $realpath = $path . '/' . $file;
     $query = file_get_contents($realpath);
@@ -35,3 +45,5 @@ foreach ($files as $file) {
     else
          echo "$file => Fail\n";
 }
+
+echo "\nCompleted\n";
